[i18nIgnore] Add new subpage handling logic (#2758)
* [i18nIgnore] Add new subpage handling logic

* Revert "[i18nIgnore] Add new subpage handling logic"

This reverts commit 5fce612aef206855e497e8e0b8fac198b6a03264.

* Refactor subpage highlighting logic

* Add missing frontmatter fields from translations

* Avoid using `import.meta.url`

---------
