+++ src/components/LeftSidebar/LeftSidebar.astro
@@ -35,7 +35,7 @@ let activeTab: 'learn' | 'api' = 'learn';
 
 // Certain pages are not in the sidebar nav, so we manually set the active tab based on other factors (e.g. Algolia page category).
 const isReference = ['Error Reference', 'Reference'].includes(
-	getPageCategory(new URL(currentPage, import.meta.url))
+	getPageCategory({ pathname: currentPage })
 );
 if (isReference) {
 	activeTab = 'api';
