+++ src/components/LeftSidebar/SidebarContent.astro
@@ -1,5 +1,6 @@
 ---
-import { getLanguageFromURL, removeSubpageSegment } from '../../util';
+import { getLanguageFromURL } from '~/util';
+import { isSubPage } from '~/util/isSubPage';
 
 export interface Props {
 	type: TabType;
@@ -44,11 +45,7 @@ const lang = getLanguageFromURL(Astro.url.pathname);
 							<a
 								href={`${Astro.site?.pathname}${lang}/${slug}/`}
 								aria-current={`${currentPageMatch.endsWith(slug) ? 'page' : 'false'}`}
-								data-current-parent={`${
-									removeSubpageSegment(currentPageMatch).endsWith(removeSubpageSegment(slug))
-										? 'true'
-										: 'false'
-								}`}
+								data-current-parent={`${isSubPage(currentPageMatch, slug) ? 'true' : 'false'}`}
 							>
 								<Fragment set:html={text} /> {isFallback && <sup class="fallback">EN</sup>}
 							</a>
