+++ src/content/docs/es/core-concepts/sharing-state.mdx
@@ -1,6 +1,8 @@
 ---
 title: Compartiendo Estado
+description: Learn how to share state across components — and frameworks! — with Nano Stores.
 i18nReady: true
+type: recipe
 ---
 
 import UIFrameworkTabs from '~/components/tabs/UIFrameworkTabs.astro'
