+++ src/content/docs/es/guides/rss.mdx
@@ -2,6 +2,7 @@
 title: RSS
 description: Introducción a RSS en Astro
 i18nReady: true
+type: recipe
 ---
 
 import PackageManagerTabs from '~/components/tabs/PackageManagerTabs.astro';
