+++ src/content/docs/fr/core-concepts/sharing-state.mdx
@@ -1,6 +1,8 @@
 ---
 title: Partage d'État
+description: Learn how to share state across components — and frameworks! — with Nano Stores.
 i18nReady: true
+type: recipe
 ---
 import UIFrameworkTabs from '~/components/tabs/UIFrameworkTabs.astro'
 import LoopingVideo from '~/components/LoopingVideo.astro'
