+++ src/content/docs/fr/guides/rss.mdx
@@ -1,6 +1,7 @@
 ---
 title: Flux RSS
 description: Une introduction aux flux RSS avec Astro.
+type: recipe
 ---
 import PackageManagerTabs from '~/components/tabs/PackageManagerTabs.astro'
 import Since from '~/components/Since.astro'
