+++ src/content/docs/ja/guides/rss.mdx
@@ -2,6 +2,7 @@
 title: RSS
 description: AstroのRSS入門
 i18nReady: true
+type: recipe
 ---
 import PackageManagerTabs from '~/components/tabs/PackageManagerTabs.astro'
