+++ src/content/docs/pt-br/core-concepts/sharing-state.mdx
@@ -1,6 +1,8 @@
 ---
 title: Compartilhamento de Estado
+description: Aprenda como compartilhar estado entre componentes — e frameworks! — com Nano Stores.
 i18nReady: true
+type: recipe
 ---
 import UIFrameworkTabs from '~/components/tabs/UIFrameworkTabs.astro';
 import LoopingVideo from '~/components/LoopingVideo.astro';
