+++ src/content/docs/pt-br/guides/rss.mdx
@@ -2,6 +2,7 @@
 title: RSS
 description: Uma introdução a RSS em Astro
 i18nReady: true
+type: recipe
 ---
 import PackageManagerTabs from '~/components/tabs/PackageManagerTabs.astro';
 import Since from '~/components/Since.astro';
