+++ src/content/docs/zh-cn/core-concepts/sharing-state.mdx
@@ -1,6 +1,8 @@
 ---
 title: 状态共享
+description: Learn how to share state across components — and frameworks! — with Nano Stores.
 i18nReady: false
+type: recipe
 ---
 import UIFrameworkTabs from '~/components/tabs/UIFrameworkTabs.astro'
 import LoopingVideo from '~/components/LoopingVideo.astro'
