+++ src/content/docs/zh-cn/guides/rss.mdx
@@ -1,6 +1,7 @@
 ---
 title: RSS
 description: Astro RSS 介绍
+type: recipe
 ---
 
 Astro 支持为博客和其他内容网站快速自动生成 RSS 摘要。更多关于 RSS 摘要的信息参见 [aboutfeeds.com](https://aboutfeeds.com/)。
